Rename imgg to img_gen
