+++ pipelex/cogt/img_gen/img_gen_handle.py
