+++ pipelex/cogt/img_gen/img_gen_job_components.py
