+++ pipelex/cogt/img_gen/img_gen_platform.py
