+++ pipelex/cogt/img_gen/img_gen_prompt.py
