+++ pipelex/cogt/inference/inference_manager.py
@@ -4,9 +4,9 @@ from typing_extensions import override
 
 from pipelex import log
 from pipelex.cogt.exceptions import InferenceManagerWorkerSetupError
-from pipelex.cogt.imgg.imgg_engine_factory import ImggEngineFactory
-from pipelex.cogt.imgg.imgg_worker_abstract import ImggWorkerAbstract
-from pipelex.cogt.imgg.imgg_worker_factory import ImggWorkerFactory
+from pipelex.cogt.img_gen.img_gen_engine_factory import ImggEngineFactory
+from pipelex.cogt.img_gen.img_gen_worker_abstract import ImggWorkerAbstract
+from pipelex.cogt.img_gen.img_gen_worker_factory import ImggWorkerFactory
 from pipelex.cogt.inference.inference_manager_protocol import InferenceManagerProtocol
 from pipelex.cogt.llm.llm_worker_abstract import LLMWorkerAbstract
 from pipelex.cogt.llm.llm_worker_factory import LLMWorkerFactory
